Remove Definite constraint on Any-iterable methods
Removing the definiteness constraint on the Any methods allows them to
be called from non-concrete type objects which produces an (arguably)
unhelpful but inoffensive result (calling .iterator on type objects will
produce an iterator that when pulled once will return the type object
again and then be exhausted).

Regardless of utility, removing the constraint on all Any-iterable
methods reduces the confusion of which methods can be called on type
objs and which cannot.