Rewrote Rat / FatRat stringification to be much faster.
Rats stringify on average twice as fast, FatRats, depending on the
size of the denominator, range from about the same to 200+ times as
fast. Also fixes a few bugs: FatRat strings very close to 1 no
longer round off to 1. Rats and FatRats no longer ever stringify
with trailing zeros in the fractional part. FatRats are now capable
of stringifying to sixteen minimum significant digits or more if
necessary to accurately express the value, but holding off for v6.c
because it isn't spec.