Remove 5to6 perlform error messages
Not enough people using perlform, so it's hard to justify support for
these error messages. Moreover, Perl 6 Form module is still
unfinished, so the recommendation to use it is somewhat unhelpful.

Note that some of the removed variables from Exception.pm6 were not
even implemented.

Also note that $^A and $^L are still illegal (like all 1-char
uppercase $^ vars), so now for these two vars it simply won't
recommend the unfinished module.

Part of #1356.