Introduce "use p5isms" pragma
When this pragma is activated in a scope, then the errors about using
Perl 5 constructs will *not* be thrown.  This does *not* do anything else.
So "defined" without any parameter will still not check $_: instead it
will just complain that calling defined without any parameters will never
work.  And *that* allows a module to export a "defined" sub that will check
$_ when called without any parameters.

I realize this is close to the 2018.05 release.  If it's too controversial
then please remove this before the release.